refactor : 패키지 생성 및 이름 변경
