+++ src/main/java/calculator/Calculator.java
@@ -1,4 +1,4 @@
-package Calculator;
+package calculator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
