+++ src/main/java/racingcar/RacingApp.java
@@ -1,0 +1,10 @@
+package racingcar;
+
+public class RacingApp {
+
+    public static void main(String[] args) {
+        Racing racing = new Racing();
+        racing.play();
+    }
+
+}
