+++ src/main/java/racingcar/util/InputScanner.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar.util;
 
 import java.util.Scanner;
