+++ src/test/java/calculator/CalculatorTest.java
@@ -1,4 +1,4 @@
-package Calculator;
+package calculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
