+++ src/test/java/racingcar/InputValidCheckerTest.java
@@ -1,5 +1,6 @@
-package RacingCar;
+package racingcar;
 
+import racingcar.util.InputValidChecker;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
